// src/superadmin/lib/api/types/student.types.ts
export interface WorkExperience {
  id: string;
  title: string;
  company: string;
  startDate: string;
  endDate?: string;
  description: string;
  current: boolean;
}